Refactor: Validate @PathVariable in controller, create separate methods for adding/removing likes and friends, and use getFilmOrThrow() and getUserOrThrow()  for 404 error handling in findFilm and findUser

+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -1,16 +1,19 @@
 package ru.yandex.practicum.filmorate.controller;
 
 import jakarta.validation.Valid;
+import jakarta.validation.constraints.Min;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
+import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.service.FilmService;
 
 import java.util.Collection;
-import java.util.Optional;
 
 @RestController
 @RequestMapping("/films")
+@Validated
 public class FilmController {
     private final FilmService filmService;
 
@@ -25,13 +28,14 @@ public class FilmController {
     }
 
     @GetMapping("/{id}")
-    public Optional<Film> findFilm(@PathVariable Long id) {
+    public Film findFilm(@PathVariable Long id) {
         return filmService.findFilm(id);
     }
 
     @GetMapping("/popular")
     public Collection<Film> getTopFilms(
-            @RequestParam(defaultValue = "10") final Long count
+            @RequestParam(defaultValue = "10")
+            @Min(value = 1, message = "The number of films must be greater than zero.") final Long count
     ) {
         return filmService.getTopFilms(count);
     }
@@ -48,11 +52,13 @@ public class FilmController {
 
     @PutMapping("/{id}/like/{userId}")
     public void addLikeToFilm(@PathVariable Long id, @PathVariable Long userId) {
+        if (userId == null || id == null) throw new ValidationException("IDs must not be null");
         filmService.addLikeToFilm(userId, id);
     }
 
     @DeleteMapping("/{id}/like/{userId}")
     public void removeLikeToFilm(@PathVariable Long id, @PathVariable Long userId) {
+        if (userId == null || id == null) throw new ValidationException("IDs must not be null");
         filmService.removeLikeToFilm(userId, id);
     }
 }
