+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -3,6 +3,7 @@ package ru.yandex.practicum.filmorate.controller;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
+import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.service.UserService;
 
@@ -25,7 +26,7 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public Optional<User> findUser(@PathVariable Long id) {
+    public User findUser(@PathVariable Long id) {
         return userService.findUser(id);
     }
 
@@ -36,6 +37,7 @@ public class UserController {
 
     @GetMapping("/{id}/friends/common/{otherId}")
     public Collection<User> getCommonFriends(@PathVariable Long id, @PathVariable Long otherId) {
+        if (id == null || otherId == null) throw new ValidationException("User IDs must not be null");
         return userService.getCommonFriends(id, otherId);
     }
 
@@ -51,11 +53,13 @@ public class UserController {
 
     @PutMapping("/{id}/friends/{friendId}")
     public User addFriend(@PathVariable Long id, @PathVariable Long friendId) {
+        if (id == null || friendId == null) throw new ValidationException("User IDs must not be null");
         return userService.addFriend(id, friendId);
     }
 
     @DeleteMapping("/{id}/friends/{friendId}")
     public User removeFriend(@PathVariable Long id, @PathVariable Long friendId) {
+        if (id == null || friendId == null) throw new ValidationException("User IDs must not be null");
         return userService.removeFriend(id, friendId);
     }
 }
